refactor: centralize `WORKER_BASE_URL` into a new global `config.ts` file.
